r12070/sofa-dev : Fix build warnings
 //partial sync (3/6 changes)//


Former-commit-id: 57df132de4db68632cdde72e5f718682209a8683

+++ modules/sofa/component/visualmodel/VisualManagerPass.cpp
@@ -106,8 +106,8 @@ void VisualManagerPass::initVisual()
     GLint viewport[4];
     glGetIntegerv(GL_VIEWPORT, viewport);
 
-    passWidth = viewport[2]*factor.getValue();
-    passHeight = viewport[3]*factor.getValue();
+    passWidth = viewport[2]*(GLint)factor.getValue();
+    passHeight = viewport[3]*(GLint)factor.getValue();
 
     fbo->init(passWidth, passHeight);
 }
